Use Linux SLAB allocator for spl_kmem_cache objects
This replaces kmem_zalloc() allocations of spl_kmem_cache objects that
are approximately 32KB in size with a Linux SLAB allocations. This
eliminates the need to supress debug warnings and makes the changes
necessary to replace the __vmalloc() in kv_alloc() with SLAB
allocations.

We also take advantage of the SLAB allocator's SLAB_HWCACHE_ALIGN flag
to reduce false sharing, which could increase performance.

Signed-off-by: Richard Yao <ryao@cs.stonybrook.edu>