Preallocate SLABs in worker threads
The SPL reimplements Solaris SLABs using Linux's virtual memory
allocator. Unfortunately, this allocator ignores gfp flags, which causes
deadlocks when used in situations where direct reclaim must be disabled.
Previous attempts at dealing with this either involvined PF_MEMALLOC or
a kernel patch, but both of these had issues. PF_MEMALLOC
had side effects that caused deadlocks while the kernel patch was not
something that the Linux kernel developers were willing to accept.

We rework the SPL SLAB allocator to allocate the next slab
in a worker thread asynchronously. If the allocation occurs
before the next slab is needed, we will obtain the next slab in O(1)
time when it is needed. If it fails to occur, we fallback to the
original behavior.

This will guarantee that the SLAB allocator will be able to service a
full slab worth of allocation requests during times of memory pressure.
This should also have the beneficial effect of causing the rest of the
kernel to feel memory pressure before the SLABs do. The early memory
pressure should encourage the kernel to make memory available in advance
of actual need by the SPL SLAB allocator.

Signed-off-by: Richard Yao <ryao@cs.stonybrook.edu>