Return immediate success from spl_cache_grow upon success
Consumers of the SPL invoke spl_kmem_cache_alloc to allocate memory from
caches. When a cache is out of memory, spl_kmem_cache_alloc() will
invoke spl_cache_refill(), which will invoke spl_cache_grow().

spl_cache_grow() will give the job of doing allocations to a task queue.
It will then call wait_event_timeout() to wait for completion. If
completion is too slow, it will time out. This gives us the option of
either telling spl_cache_refill() that we succeeded and there should be
memory or we did not succeed. Currently, we return ENOMEM
unconditionally, which forces us to do another iteration to figure out
that we really should have returned 0 for success. This is
unnecessary, so lets return 0 immediately upon success.

Signed-off-by: Richard Yao <ryao@gentoo.org>