Do not defer emergency allocations
A concept of an emergency allocation was introduced last year to permit
allocations that took excessive periods of time to switch to physical
memory to make forward progress. Emergency allocations currently occur
by setting a flag, telling kmem_cache_alloc() to try again and then
doing the allocation. This is unnecessary, so we restructure the code
to do the emergency allocations immediately.

Signed-off-by: Richard Yao <ryao@gentoo.org>