Kill taskq code
This disables removes the taskq code from the SPL in favor of a port of
the Illumos taskq code that will sit on top of the SPL with ZFS. It also
removes the corresponding SPLAT tests. A few splat tests depended on the
taskq code, so those have been #ifdefed out until they can be rewritten
to use the Linux workqueue interface.

Signed-off-by: Richard Yao <ryao@gentoo.org>