Add sys/klwp.h and sys/class.h for better Solaris compatibility
The Illumos taskq code must include sys/condvar.h. On Illumos, this is
done by including sys/thread.h, which includes sys/klwp.h, which
includes sys/condvar.h. To avoid the need to change code, we add
sys/klwp.h, which includes sys/condvar.h and make sys/thread.h include
it. There are more things that sys/klwp.h is intended to include, but we
only include what we need for now. This should be revisited in the
future.

We also add a dummy sys/class.h for the purposes of making the taskq
code happy.

Signed-off-by: Richard Yao <ryao@gentoo.org>