Implement cv_reltimedwait()
The enum time_res_t and its comment were taken from illumos-gate. The
enum should not be subject to copyright as per the Google versus Oracle
lawsuit and general precedent. The enum should not be subject to
copyright of any of two reasons that just happen to both apply. The
first is that it is trivial to recreate from public documentation, so it
could be derived independently. The second is that headers needed for
compatibility are universally considered to be exempt from copyright,
which was upheld in the Google versus Oracle lawsuit. The comment might
be arugably subject to copyright, but it is so small that its
duplication should fall under fall use.

Signed-off-by: Richard Yao <ryao@gentoo.org>