KM_SLEEP kmem_cache allocations should not block high priority allocations
If we have a cache that has no free objects, two concurrent allocations
occur from a cache where one is KM_SLEEP and another is KM_PUSHPAGE in a
the sync task context (for example), the KM_SLEEP begins growing the
cache, the KM_PUSHPAGE allocation is silently converted to a KM_SLEEP
allocation. We have been changing allocations from KM_SLEEP to
KM_PUSHPAGE to try to avoid deadlocks. If there is merit to that, then
we should not allow races to randomly convert KM_PUSHPAGE allocations
into KM_SLEEP allocations.

Taking advantage of PF_FSTRANS will allow us to switch all allocations
that we changed to KM_PUSHPAGE back to KM_SLEEP because KM_SLEEP
allocations in key callpaths will automatically become KM_PUSHPAGE
allocations. That will make avoiding the potential for a KM_PUSHPAGE
allocation to be silently converted into a KM_SLEEP allocation even more
important because there will be increased opportunity for it. It might
seem better to let everything use KM_PUSHPAGE, but that is not the case
because allowing us to block in all allocations that are not necessary
for forward progress will reduce our burden under memory pressure, which
will ultimately improve system throughput.

Signed-off-by: Richard Yao <ryao@gentoo.org>